// SDKs/twilio.com/twilio_preview/1.36.0/typescript/src/sdk/models/shared/previewhostednumbershostednumberorder.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { HostedNumberOrderEnumStatusEnum } from "./hostednumberorderenumstatusenum";
import { HostedNumberOrderEnumVerificationTypeEnum } from "./hostednumberorderenumverificationtypeenum";


// PreviewHostedNumbersHostedNumberOrderCapabilities
/** 
 * A mapping of phone number capabilities.
**/
export class PreviewHostedNumbersHostedNumberOrderCapabilities extends SpeakeasyBase {
  @Metadata({ data: "json, name=fax" })
  fax?: boolean;

  @Metadata({ data: "json, name=mms" })
  mms?: boolean;

  @Metadata({ data: "json, name=sms" })
  sms?: boolean;

  @Metadata({ data: "json, name=voice" })
  voice?: boolean;
}


export class PreviewHostedNumbersHostedNumberOrder extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=address_sid" })
  addressSid?: string;

  @Metadata({ data: "json, name=call_delay" })
  callDelay?: number;

  @Metadata({ data: "json, name=capabilities" })
  capabilities?: PreviewHostedNumbersHostedNumberOrderCapabilities;

  @Metadata({ data: "json, name=cc_emails" })
  ccEmails?: string[];

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=email" })
  email?: string;

  @Metadata({ data: "json, name=extension" })
  extension?: string;

  @Metadata({ data: "json, name=failure_reason" })
  failureReason?: string;

  @Metadata({ data: "json, name=friendly_name" })
  friendlyName?: string;

  @Metadata({ data: "json, name=incoming_phone_number_sid" })
  incomingPhoneNumberSid?: string;

  @Metadata({ data: "json, name=phone_number" })
  phoneNumber?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=signing_document_sid" })
  signingDocumentSid?: string;

  @Metadata({ data: "json, name=status" })
  status?: HostedNumberOrderEnumStatusEnum;

  @Metadata({ data: "json, name=unique_name" })
  uniqueName?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;

  @Metadata({ data: "json, name=verification_attempts" })
  verificationAttempts?: number;

  @Metadata({ data: "json, name=verification_call_sids" })
  verificationCallSids?: string[];

  @Metadata({ data: "json, name=verification_code" })
  verificationCode?: string;

  @Metadata({ data: "json, name=verification_document_sid" })
  verificationDocumentSid?: string;

  @Metadata({ data: "json, name=verification_type" })
  verificationType?: HostedNumberOrderEnumVerificationTypeEnum;
}